Add Petpooja attendance import → Working Days (with half-days)
New "⬆ Attendance" tool on the Payroll tab: upload a CSV with Name + Present
Days (Present = Full + ½·Half, as aggregated from a Petpooja Attendance Master),
match each row to an employee (review-before-apply, reusing the bank-import
matcher), and set that employee's Working Days for the selected month.

Working days may now be fractional to represent half-days: dropped the .int()
constraint on the payroll adjust wd (SQLite stores 22.5 fine), and the on-screen
Work Days input steps by 0.5. Verified: importing sets wd correctly and earned =
salary × wd / baseDays (incl. a 22.5-day half-day case). Tests 30 green.

Note: the app imports a CSV; the Petpooja report is a PDF, so for now the PDF is
converted to this CSV offline. Direct in-app PDF upload can follow.

+++ src/routes/payroll.js
@@ -34,7 +34,8 @@ router.get("/:mk", (req, res) => {
 // would let z.coerce.number() turn null into 0, making it impossible to clear
 // an override (it would stay stuck at 0).
 const AdjustSchema = z.object({
-  wd: z.coerce.number().int().min(0).max(366).nullable().optional(),
+  // Working days may be fractional (a half-day worked = 0.5), e.g. from attendance.
+  wd: z.coerce.number().min(0).max(366).nullable().optional(),
   bonus: z.coerce.number().int().min(0).max(100_000_000).optional(),
   ded: z.coerce.number().int().min(0).max(100_000_000).optional(),
   adv: z.coerce.number().int().min(0).max(100_000_000).nullable().optional(),
